Move `app <app-name> get commits` to `get commits <app-name>`
