+++ cmd/gitops/get/cmd.go
@@ -5,6 +5,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/app"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/clusters"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/get/commits"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/credentials"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/templates"
 )
@@ -17,6 +18,9 @@ func GetCommand(endpoint *string, client *resty.Client) *cobra.Command {
 # Get all applications under gitops control
 gitops get apps
 
+# Get last 10 commits for an application
+gitops get commits <app-name>
+
 # Get all CAPI templates
 gitops get templates
 
@@ -28,6 +32,7 @@ gitops get clusters`,
 	}
 
 	cmd.AddCommand(app.Cmd)
+	cmd.AddCommand(commits.Cmd)
 	cmd.AddCommand(templates.TemplateCommand(endpoint, client))
 	cmd.AddCommand(credentials.CredentialCommand(endpoint, client))
 	cmd.AddCommand(clusters.ClusterCommand(endpoint, client))
